+++ drivers/pds/lib/utils.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"math/rand"
 	"net/http"
+	"net/url"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -21,6 +22,7 @@ import (
 	"github.com/portworx/sched-ops/k8s/apps"
 	"github.com/portworx/sched-ops/k8s/core"
 	pdsapi "github.com/portworx/torpedo/drivers/pds/api"
+	pdscontrolplane "github.com/portworx/torpedo/drivers/pds/controlplane"
 	tc "github.com/portworx/torpedo/drivers/pds/targetcluster"
 	v1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -188,14 +190,14 @@ var (
 	isBuildAvailable                      bool
 	currentReplicas                       int32
 	deploymentTargetID, storageTemplateID string
-	accountID                             string
-	tenantID                              string
-	projectID                             string
 	resourceTemplateID                    string
 	appConfigTemplateID                   string
 	versionID                             string
 	imageID                               string
 	serviceAccId                          string
+	accountID                             string
+	projectID                             string
+	tenantID                              string
 	istargetclusterAvailable              bool
 	isAccountAvailable                    bool
 	isStorageTemplateAvailable            bool
@@ -634,7 +636,7 @@ func GetVersionsImage(dsVersion string, dsBuild string, dataServiceID string) (s
 		}
 	}
 	if !(isVersionAvailable && isBuildAvailable) {
-		log.Errorf("Version/Build passed is not available")
+		return "", "", nil, fmt.Errorf("version/build passed is not available")
 	}
 	return versionID, imageID, dataServiceVersionBuildMap, nil
 }
@@ -1038,6 +1040,81 @@ func IsReachable(url string) (bool, error) {
 	return true, nil
 }
 
+// SetupPDSTest returns few params required to run the test
+func SetupPDSTest(ControlPlaneURL, ClusterType, AccountName, TenantName, ProjectName string) (string, string, string, string, string, error) {
+	var err error
+	apiConf := pds.NewConfiguration()
+	endpointURL, err := url.Parse(ControlPlaneURL)
+	if err != nil {
+		return "", "", "", "", "", err
+	}
+	apiConf.Host = endpointURL.Host
+	apiConf.Scheme = endpointURL.Scheme
+
+	apiClient = pds.NewAPIClient(apiConf)
+	components = pdsapi.NewComponents(apiClient)
+	controlplane := pdscontrolplane.NewControlPlane(ControlPlaneURL, components)
+
+	if strings.EqualFold(ClusterType, "onprem") || strings.EqualFold(ClusterType, "ocp") {
+		serviceType = "ClusterIP"
+	}
+	log.InfoD("Deployment service type %s", serviceType)
+
+	acc := components.Account
+	accounts, err := acc.GetAccountsList()
+	if err != nil {
+		return "", "", "", "", "", err
+	}
+
+	isAccountAvailable = false
+	for i := 0; i < len(accounts); i++ {
+		log.InfoD("Account Name: %v", accounts[i].GetName())
+		if accounts[i].GetName() == AccountName {
+			isAccountAvailable = true
+			accountID = accounts[i].GetId()
+		}
+	}
+	if !isAccountAvailable {
+		log.Fatalf("Account %v is not available", AccountName)
+	}
+	log.InfoD("Account Detail- Name: %s, UUID: %s ", AccountName, accountID)
+	tnts := components.Tenant
+	tenants, _ := tnts.GetTenantsList(accountID)
+	for _, tenant := range tenants {
+		if tenant.GetName() == TenantName {
+			tenantID = tenant.GetId()
+		}
+
+	}
+	log.InfoD("Tenant Details- Name: %s, UUID: %s ", TenantName, tenantID)
+	dnsZone, err := controlplane.GetDNSZone(tenantID)
+	if err != nil {
+		return "", "", "", "", "", err
+	}
+	log.InfoD("DNSZone: %s, tenantName: %s, accountName: %s", dnsZone, TenantName, AccountName)
+	projcts := components.Project
+	projects, _ := projcts.GetprojectsList(tenantID)
+	for _, project := range projects {
+		if project.GetName() == ProjectName {
+			projectID = project.GetId()
+		}
+	}
+	log.InfoD("Project Details- Name: %s, UUID: %s ", ProjectName, projectID)
+
+	ns, err = k8sCore.GetNamespace("kube-system")
+	if err != nil {
+		return "", "", "", "", "", err
+	}
+	clusterID := string(ns.GetObjectMeta().GetUID())
+	if len(clusterID) > 0 {
+		log.InfoD("clusterID %v", clusterID)
+	} else {
+		return "", "", "", "", "", fmt.Errorf("unable to get the clusterID")
+	}
+
+	return tenantID, dnsZone, projectID, serviceType, clusterID, err
+}
+
 //RegisterClusterToControlPlane checks and registers the given target cluster to the controlplane
 func RegisterClusterToControlPlane(controlPlaneUrl, tenantId, clusterType string) error {
 	log.InfoD("Test control plane url connectivity.")
@@ -1051,7 +1128,7 @@ func RegisterClusterToControlPlane(controlPlaneUrl, tenantId, clusterType string
 		return fmt.Errorf("error while getting helm version - %v", helmChartversion)
 	}
 
-	log.Info("listing service account")
+	log.InfoD("Listing service account")
 	listServiceAccounts, err := components.ServiceAccount.ListServiceAccounts(tenantId)
 	if err != nil {
 		return err
@@ -1063,7 +1140,7 @@ func RegisterClusterToControlPlane(controlPlaneUrl, tenantId, clusterType string
 		}
 	}
 
-	log.Info("getting service account token")
+	log.InfoD("Getting service account token")
 	serviceAccToken, err := components.ServiceAccount.GetServiceAccountToken(serviceAccId)
 	if err != nil {
 		return err
@@ -1072,7 +1149,7 @@ func RegisterClusterToControlPlane(controlPlaneUrl, tenantId, clusterType string
 
 	ctx := GetAndExpectStringEnvVar("TARGET_KUBECONFIG")
 	target := tc.NewTargetCluster(ctx)
-	err = target.RegisterToControlPlane(controlPlaneUrl, helmChartversion, bearerToken, tenantID, clusterType)
+	err = target.RegisterToControlPlane(controlPlaneUrl, helmChartversion, bearerToken, tenantId, clusterType)
 	if err != nil {
 		return fmt.Errorf("target cluster registeration failed with the error: %v", err)
 	}
