+++ tests/pds/common.go
@@ -1,14 +1,9 @@
 package tests
 
 import (
-	"fmt"
-	"net/url"
-	"strings"
-
 	pds "github.com/portworx/pds-api-go-client/pds/v1alpha1"
 	"github.com/portworx/sched-ops/k8s/core"
 	pdsapi "github.com/portworx/torpedo/drivers/pds/api"
-	pdscontrolplane "github.com/portworx/torpedo/drivers/pds/controlplane"
 	pdslib "github.com/portworx/torpedo/drivers/pds/lib"
 	"github.com/portworx/torpedo/pkg/aetosutil"
 	"github.com/portworx/torpedo/pkg/log"
@@ -140,79 +135,4 @@ func RunWorkloads(params pdslib.WorkloadGenerationParams, ds PDSDataService, dep
 
 	return pod, dep, err
 
-}
-
-// SetupPDSTest returns few params required to run the test
-func SetupPDSTest(ControlPlaneURL, ClusterType, AccountName, TenantName, ProjectName string) (string, string, string, string, string, error) {
-	var err error
-	apiConf := pds.NewConfiguration()
-	endpointURL, err := url.Parse(ControlPlaneURL)
-	if err != nil {
-		return "", "", "", "", "", err
-	}
-	apiConf.Host = endpointURL.Host
-	apiConf.Scheme = endpointURL.Scheme
-
-	apiClient = pds.NewAPIClient(apiConf)
-	components = pdsapi.NewComponents(apiClient)
-	controlplane := pdscontrolplane.NewControlPlane(ControlPlaneURL, components)
-
-	if strings.EqualFold(ClusterType, "onprem") || strings.EqualFold(ClusterType, "ocp") {
-		serviceType = "ClusterIP"
-	}
-	log.InfoD("Deployment service type %s", serviceType)
-
-	acc := components.Account
-	accounts, err := acc.GetAccountsList()
-	if err != nil {
-		return "", "", "", "", "", err
-	}
-
-	isAccountAvailable = false
-	for i := 0; i < len(accounts); i++ {
-		log.InfoD("Account Name: %v", accounts[i].GetName())
-		if accounts[i].GetName() == AccountName {
-			isAccountAvailable = true
-			accountID = accounts[i].GetId()
-		}
-	}
-	if !isAccountAvailable {
-		log.Fatalf("Account %v is not available", AccountName)
-	}
-	log.InfoD("Account Detail- Name: %s, UUID: %s ", AccountName, accountID)
-	tnts := components.Tenant
-	tenants, _ := tnts.GetTenantsList(accountID)
-	for _, tenant := range tenants {
-		if tenant.GetName() == TenantName {
-			tenantID = tenant.GetId()
-		}
-
-	}
-	log.InfoD("Tenant Details- Name: %s, UUID: %s ", TenantName, tenantID)
-	dnsZone, err := controlplane.GetDNSZone(tenantID)
-	if err != nil {
-		return "", "", "", "", "", err
-	}
-	log.InfoD("DNSZone: %s, tenantName: %s, accountName: %s", dnsZone, TenantName, AccountName)
-	projcts := components.Project
-	projects, _ := projcts.GetprojectsList(tenantID)
-	for _, project := range projects {
-		if project.GetName() == ProjectName {
-			projectID = project.GetId()
-		}
-	}
-	log.InfoD("Project Details- Name: %s, UUID: %s ", ProjectName, projectID)
-
-	ns, err = k8sCore.GetNamespace("kube-system")
-	if err != nil {
-		return "", "", "", "", "", err
-	}
-	clusterID := string(ns.GetObjectMeta().GetUID())
-	if len(clusterID) > 0 {
-		log.InfoD("clusterID %v", clusterID)
-	} else {
-		return "", "", "", "", "", fmt.Errorf("unable to get the clusterID")
-	}
-
-	return tenantID, dnsZone, projectID, serviceType, clusterID, err
 }
